VM: the aging_policy and nursery_policy classes are so small that they can be in the same cpp file with the methods that use them
