+++ vm/nursery_collector.cpp
@@ -2,6 +2,20 @@
 
 namespace factor {
 
+struct nursery_policy {
+  factor_vm* parent;
+
+  explicit nursery_policy(factor_vm* parent) : parent(parent) {}
+
+  bool should_copy_p(object* obj) {
+    return parent->data->nursery->contains_p(obj);
+  }
+
+  void promoted_object(object* obj) {}
+
+  void visited_object(object* obj) {}
+};
+
 void factor_vm::collect_nursery() {
   /* Copy live objects from the nursery (as determined by the root set and
      marked cards in aging and tenured) to aging space. */
